Clean up what should be visible
