+++ molot/extras.py
@@ -2,13 +2,7 @@ import subprocess
 import sys
 from typing import Optional
 
-
-class ReturnCodeError(Exception):
-    """Error for failed shell commands in piped mode."""
-
-    def __init__(self, code: int, output: str):
-        self.code = code
-        self.output = output
+from .errors import ShellReturnError
 
 
 def shell(command: str, piped: bool = False, sensitive: bool = False) -> Optional[str]:
@@ -20,7 +14,7 @@ def shell(command: str, piped: bool = False, sensitive: bool = False) -> Optiona
         sensitive (bool, optional): Control whether shell command is hidden from logs. Defaults to False.
 
     Raises:
-        ReturnCodeError: Raised when command failed in piped mode.
+        ShellReturnError: Raised when command failed in piped mode.
 
     Returns:
         Optional[str]: Returns string output when piped, otherwise nothing.
@@ -41,7 +35,7 @@ def shell(command: str, piped: bool = False, sensitive: bool = False) -> Optiona
 
     if p.returncode != 0:
         if piped:
-            raise ReturnCodeError(p.returncode, err.decode("utf-8"))
+            raise ShellReturnError(p.returncode, err.decode("utf-8"))
         sys.exit(p.returncode)
 
     return out.decode("utf-8") if piped else None
