+++ tests/__init__.py
@@ -1,6 +1,6 @@
 import logging
 
-from molot.package import setup_logging
+from molot._package import setup_logging
 
 from .test_core import *
 from .test_extras import *
